Move overlay and basemap controls to informationpanel
